+++ edgegrid-signer-apache-http-client/src/main/java/com/akamai/edgegrid/signer/apachehttpclient/ApacheHttpClientEdgeGridRoutePlanner.java
@@ -43,7 +43,17 @@ public class ApacheHttpClientEdgeGridRoutePlanner extends SystemDefaultRoutePlan
      * @param clientCredential a {@link ClientCredential}
      */
     public ApacheHttpClientEdgeGridRoutePlanner(ClientCredential clientCredential) {
-        super(ProxySelector.getDefault());
+        this(clientCredential, ProxySelector.getDefault());
+    }
+
+    /**
+     * Creates an EdgeGrid route planner using {@link ClientCredential}.
+     *
+     * @param clientCredential a {@link ClientCredential}
+     * @param proxySelector a {@link ProxySelector}
+     */
+    public ApacheHttpClientEdgeGridRoutePlanner(ClientCredential clientCredential, ProxySelector proxySelector) {
+        super(proxySelector);
         this.binding = new ApacheHttpClientEdgeGridRequestSigner(clientCredential);
     }
 
@@ -53,7 +63,17 @@ public class ApacheHttpClientEdgeGridRoutePlanner extends SystemDefaultRoutePlan
      * @param clientCredentialProvider a {@link ClientCredentialProvider}
      */
     public ApacheHttpClientEdgeGridRoutePlanner(ClientCredentialProvider clientCredentialProvider) {
-        super(ProxySelector.getDefault());
+        this(clientCredentialProvider, ProxySelector.getDefault());
+    }
+
+    /**
+     * Creates an EdgeGrid route planner using {@link ClientCredentialProvider}.
+     *
+     * @param clientCredentialProvider a {@link ClientCredentialProvider}
+     * @param proxySelector a {@link ProxySelector}
+     */
+    public ApacheHttpClientEdgeGridRoutePlanner(ClientCredentialProvider clientCredentialProvider, ProxySelector proxySelector) {
+        super(proxySelector);
         this.binding = new ApacheHttpClientEdgeGridRequestSigner(clientCredentialProvider);
     }